ndsp sound

+++ arm11/source/MixerHardware.h
@@ -1,5 +1,5 @@
 #pragma once
-
+#include <3ds/types.h>
 class MixerHardware {
 public:
 	MixerHardware();
@@ -7,7 +7,7 @@ public:
 	virtual void flush() = 0;
 	virtual void update(int *pAudioData, int count) = 0;
 	virtual void update(short *pAudioData, int count) = 0;
-	virtual int samplepos() = 0;
+	virtual u64 samplepos() = 0;
 };
 
 inline MixerHardware::MixerHardware() {
